fix warnings add config configs

+++ src/string_helpers.hpp
@@ -2,9 +2,8 @@
 #include <string>
 #include <Windows.h>
 
-
-std::string utf8_encode(const std::wstring&);
-std::wstring utf8_decode(const std::string&);
+std::string utf8_encode(const std::wstring &);
+std::wstring utf8_decode(const std::string &);
 // https://stackoverflow.com/a/54491532
 template <typename TChar, typename TStringGetterFunc>
 std::basic_string<TChar> GetStringFromWindowsApi(TStringGetterFunc stringGetter, int initialSize = 0)
@@ -17,7 +16,7 @@ std::basic_string<TChar> GetStringFromWindowsApi(TStringGetterFunc stringGetter,
     std::basic_string<TChar> result(initialSize, 0);
     for (;;)
     {
-        auto length = stringGetter(&result[0], result.length());
+        auto length = stringGetter(&result[0], static_cast<int>(result.length()));
         if (length == 0)
         {
             return std::basic_string<TChar>();
